gamepad count() to size_t

+++ src/plugins/input_system.h
@@ -154,7 +154,7 @@ struct input : developer::Plugin {
 
   struct ProvidesMaxGamepadID : public BaseComponent {
     int max_gamepad_available = 0;
-    int count() const { return max_gamepad_available + 1; }
+    size_t count() const { return (size_t)max_gamepad_available + 1; }
   };
 
   template <typename Action>
